+++ http/https.go
@@ -2,7 +2,6 @@ package wiredhttps
 
 import (
 	"crypto/tls"
-	"encoding/json"
 	"fmt"
 	"io"
 	"net"
@@ -10,11 +9,8 @@ import (
 	"strings"
 	"sync"
 	"time"
-	"wiredshield/http/routes"
 	"wiredshield/modules/db"
 	"wiredshield/modules/env"
-	"wiredshield/modules/logging"
-	"wiredshield/modules/whois"
 	"wiredshield/services"
 
 	_ "github.com/lib/pq"
@@ -22,20 +18,6 @@ import (
 )
 
 var (
-	clientPool = sync.Pool{
-		New: func() interface{} {
-			return &fasthttp.Client{
-				ReadTimeout:     60 * time.Second,
-				WriteTimeout:    60 * time.Second,
-				MaxConnsPerHost: 1024 * 256,
-				Dial: (&fasthttp.TCPDialer{
-					Concurrency:      1024 * 256,
-					DNSCacheDuration: 5 * time.Minute,
-				}).Dial,
-			}
-		},
-	}
-
 	service       *services.Service
 	certCache     sync.Map
 	certLoadMutex sync.RWMutex
@@ -50,62 +32,11 @@ func Prepare(_service *services.Service) func() {
 	service = _service
 
 	return func() {
-		binding := env.GetEnv("HTTP_BINDING", "")
-		addr := binding + ":" + env.GetEnv("HTTP_PORT", "443")
-
-		service.InfoLog("Starting HTTPS proxy on " + addr)
-		service.OnlineSince = time.Now().Unix()
-
-		server := &fasthttp.Server{
-			Concurrency:   1024 * 256,
-			Handler:       proxyHandler,
-			Name:          "wiredshield",
-			MaxConnsPerIP: 1024 * 256,
-			TLSConfig: &tls.Config{
-				NextProtos:               []string{"http/1.1"},
-				MinVersion:               tls.VersionTLS10,
-				MaxVersion:               tls.VersionTLS13,
-				GetCertificate:           getCertificateForDomain,
-				InsecureSkipVerify:       false,
-				ClientCAs:                nil,
-				PreferServerCipherSuites: true,
-				GetConfigForClient: func(hello *tls.ClientHelloInfo) (*tls.Config, error) {
-					return &tls.Config{
-						NextProtos:               []string{"http/1.1"},
-						MinVersion:               tls.VersionTLS10,
-						MaxVersion:               tls.VersionTLS13,
-						GetCertificate:           getCertificateForDomain,
-						InsecureSkipVerify:       false,
-						ClientCAs:                nil,
-						PreferServerCipherSuites: true,
-						CipherSuites: []uint16{
-							tls.TLS_AES_128_GCM_SHA256,
-							tls.TLS_AES_256_GCM_SHA384,
-							tls.TLS_CHACHA20_POLY1305_SHA256,
-							tls.TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,
-							tls.TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
-							tls.TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305,
-							tls.TLS_RSA_WITH_AES_128_CBC_SHA,
-							tls.TLS_RSA_WITH_AES_256_CBC_SHA,
-							tls.TLS_RSA_WITH_AES_128_GCM_SHA256,
-							tls.TLS_RSA_WITH_AES_128_CBC_SHA,
-							tls.TLS_RSA_WITH_AES_256_CBC_SHA,
-							tls.TLS_RSA_WITH_3DES_EDE_CBC_SHA,
-						},
-						CurvePreferences: []tls.CurveID{
-							tls.X25519,
-							tls.CurveP256,
-							tls.CurveP384,
-						},
-					}, nil
-				},
-			},
-		}
-
+		// logging
 		go processRequestLogs()
 
+		// http redirect
 		go func() {
-			httpAddr := binding + ":80"
 			service.InfoLog("Starting HTTP redirect server on " + httpAddr)
 			httpServer := &http.Server{
 				Addr:    httpAddr,
@@ -118,7 +49,10 @@ func Prepare(_service *services.Service) func() {
 			}
 		}()
 
-		ln, err := net.Listen("tcp", addr)
+		// https logic
+		service.InfoLog("Starting HTTPS proxy on " + httpsAddr)
+		service.OnlineSince = time.Now().Unix()
+		ln, err := net.Listen("tcp", httpsAddr)
 		if err != nil {
 			service.FatalLog(fmt.Sprintf("Error creating listener: %v", err))
 			return
@@ -132,23 +66,15 @@ func Prepare(_service *services.Service) func() {
 	}
 }
 
-func redirectToHTTPS(w http.ResponseWriter, r *http.Request) {
-	target := "https://" + r.Host + r.URL.Path
-	if len(r.URL.RawQuery) > 0 {
-		target += "?" + r.URL.RawQuery
-	}
-
-	http.Redirect(w, r, target, http.StatusMovedPermanently)
-}
-
-func proxyHandler(ctx *fasthttp.RequestCtx) {
+func httpsProxyHandler(ctx *fasthttp.RequestCtx) {
 	defer func() {
 		if err := recover(); err != nil {
-			service.ErrorLog(fmt.Sprintf("recovered from panic in proxyHandler: %v", err))
+			service.ErrorLog(fmt.Sprintf("recovered from panic in httpsProxyHandler: %v", err))
 			ctx.Error("Internal Server Error (Backend Panic)", fasthttp.StatusInternalServerError)
 		}
 	}()
 
+	// internal routes
 	if strings.HasPrefix(string(ctx.Path()), "/.wiredshield/") {
 		handleWiredShieldEndpoints(ctx)
 		return
@@ -218,146 +144,13 @@ func proxyHandler(ctx *fasthttp.RequestCtx) {
 	logRequest(ctx, resp, timeStart, resp.StatusCode(), requestSize, getResponseSize(ctx, resp))
 }
 
-func handleWiredShieldEndpoints(ctx *fasthttp.RequestCtx) {
-	switch string(ctx.Path()) {
-	case "/.wiredshield/proxy-auth":
-		routes.ProxyAuth(ctx)
-		return
-	case "/.wiredshield/info":
-		routes.Info(ctx)
-		return
-	default:
-		ctx.Error("not found", fasthttp.StatusNotFound)
-	}
-}
-
-// both methods still inaccurate, will fix at a later time
-func getRequestSize(ctx *fasthttp.RequestCtx) int64 {
-	totalSize := int64(0)
-
-	totalSize += int64(len(ctx.Method()))
-	url := ctx.URI().String()
-	totalSize += int64(len(url))
-
-	ctx.Request.Header.VisitAll(func(key, value []byte) {
-		totalSize += int64(len(key) + len(value))
-	})
-
-	totalSize += int64(len(ctx.Request.Body()))
-
-	return totalSize
-}
-
-func getResponseSize(ctx *fasthttp.RequestCtx, resp *fasthttp.Response) int64 {
-	totalSize := int64(0)
-
-	totalSize += int64(len(fmt.Sprintf("%d", resp.StatusCode())))
-
-	resp.Header.VisitAll(func(key, value []byte) {
-		totalSize += int64(len(key) + len(value))
-	})
-
-	totalSize += int64(len(resp.Body()))
-
-	return totalSize
-}
-
-func logRequest(ctx *fasthttp.RequestCtx, resp *fasthttp.Response, timeStart time.Time, internalCode int, requestSize, responseSize int64) {
-	reqHeadersMap := make(map[string]string)
-	ctx.Request.Header.VisitAll(func(key, value []byte) {
-		reqHeadersMap[string(key)] = string(value)
-	})
-	reqHeaders, _ := json.Marshal(reqHeadersMap)
-
-	respHeadersMap := make(map[string]string)
-	ctx.Response.Header.VisitAll(func(key, value []byte) {
-		respHeadersMap[string(key)] = string(value)
-	})
-	respHeaders, _ := json.Marshal(respHeadersMap)
-
-	responseStatusOrigin := 0
-	if resp != nil {
-		responseStatusOrigin = resp.StatusCode()
+func redirectToHTTPS(w http.ResponseWriter, r *http.Request) {
+	target := "https://" + r.Host + r.URL.Path
+	if len(r.URL.RawQuery) > 0 {
+		target += "?" + r.URL.RawQuery
 	}
 
-	ip := getIp(ctx)
-	country, err := whois.GetCountry(ip)
-	if err != nil {
-		service.ErrorLog(fmt.Sprintf("failed to get country for IP %s: %v", ip, err))
-	}
-
-	logging.RequestLogsChannel <- &logging.HTTPRequestLog{
-		RequestTime:          timeStart.UnixMilli(),
-		ClientIP:             ip,
-		Method:               string(ctx.Method()),
-		Host:                 string(ctx.Host()),
-		Path:                 string(ctx.Path()),
-		QueryParams:          queryParamString(string(ctx.QueryArgs().String())),
-		RequestHeaders:       json.RawMessage(reqHeaders),
-		ResponseHeaders:      json.RawMessage(respHeaders),
-		ResponseStatusOrigin: responseStatusOrigin,
-		ResponseStatusProxy: func() int {
-			if internalCode != 0 {
-				return internalCode
-			}
-
-			return resp.StatusCode()
-		}(),
-		ResponseTime:       time.Since(timeStart).Milliseconds(),
-		TLSVersion:         tlsVersionToString(ctx.TLSConnectionState().Version),
-		RequestSize:        requestSize,
-		ResponseSize:       responseSize,
-		RequestHTTPVersion: string(ctx.Request.Header.Protocol()),
-		ClientCountry:      country,
-	}
-}
-
-type QueryParams map[string]string
-
-func queryParamString(query string) json.RawMessage {
-	params := make(QueryParams)
-	for _, pair := range strings.Split(query, "&") {
-		parts := strings.Split(pair, "=")
-		if len(parts) != 2 {
-			continue
-		}
-
-		params[parts[0]] = parts[1]
-	}
-
-	data, _ := json.Marshal(params)
-	return data
-}
-
-func tlsVersionToString(version uint16) string {
-	switch version {
-	case tls.VersionTLS10:
-		return "TLS 1.0"
-	case tls.VersionTLS11:
-		return "TLS 1.1"
-	case tls.VersionTLS12:
-		return "TLS 1.2"
-	case tls.VersionTLS13:
-		return "TLS 1.3"
-	default:
-		return "Unknown"
-	}
-}
-
-func processRequestLogs() {
-	logWorkers := 512
-	for i := 0; i < logWorkers; i++ {
-		go func() {
-			for log := range logging.RequestLogsChannel {
-				logs := log.CollectAdditionalLogs()
-				if len(logs) > 0 {
-					if err := log.BatchInsert(logs); err != nil {
-						service.ErrorLog(fmt.Sprintf("Failed to insert logs: %v", err))
-					}
-				}
-			}
-		}()
-	}
+	http.Redirect(w, r, target, http.StatusMovedPermanently)
 }
 
 func getCertificateForDomain(hello *tls.ClientHelloInfo) (*tls.Certificate, error) {
